digital-fte/requirements.txt
watchdog
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
schedule
psutil